// frontend/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        NoteCreate: {
            /** Title */
            title: string;
            /** Content */
            content: string;
            /** Folder Id */
            folder_id?: number | null;
        };
        /** NoteRead */
        NoteRead: {
            /** Id */
            id: number;
            /** Title */
            title: string;
            /** Content */
            content: string;
            /** Folder Id */
            folder_id?: number | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
            /**
             * Tags
             * @default []
             */
            tags: components["schemas"]["TagRead"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}